Move files to UnitTests direction
